Add importStyles config options
Having styles import automatically when a component
is mounted is nice but creates unwanted artifacts.

Alternatively, they can be added to React's public/index.html
once for all pages in which case this config option allows
the automatic imports to be disabled.

The option is true by default.

There is currently no way to set it via command line arguments.