render proxies continuously during React render
Previously, Appfairy would render all proxies of a view in one go.
The tree returned by the render function of a view always had all
its proxies resolved.

To split this process into multiple components so that they can be
updated independently, a child view was used.

Now, the render function of a view will instead return the page
with <Hatch> components at every point where a proxy needs to be
resolved.

This allows the page render to be split without the need to
create child views.

The af-view feature has therefore been removed, every page
consists of a single view only.

The update also includes a new syntax for proxies.

<MyView>
  <Render>
    <Proxy sock={sock.home} merge>
      <a href="/" />
    </Proxy>
  </Render>
</MyView>

It is more verbose but also more readable and flexible with
a number of new features.

The sock enums are now exported as pageNameSock (indexSock,
aboutUsSock, etc.) and are included in index.js files.

They use a simplified structure and have comments right on
the enum, fulfilling the role of the view docstring which
has therefore been removed.